[WinEH] Don't assume endFunction is called while in .text
Jump table emission can switch to .rdata before
WinException::endFunction gets called. Just remember the appropriate
text section we started in and reset back to it when we end the
function. We were already switching sections back from .xdata anyway.

Fixes the first problem in PR31488, so that now COFF switch tables can
live in .rdata if we want them to.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@290678 91177308-0d34-0410-b5e6-96231b3b80d8